split out AST to keep it cleaner and add some quick utility types
